use sequence_staged for alter cluster ddl
