+++ src/adapter/src/coord/sequencer/inner.rs
@@ -115,6 +115,7 @@ use crate::session::{
 use crate::util::{viewable_variables, ClientTransmitter, ResultExt};
 use crate::{guard_write_critical_section, PeekResponseUnary, ReadHolds};
 
+mod cluster;
 mod create_index;
 mod create_materialized_view;
 mod create_view;
