+++ src/adapter/src/coord/sequencer/inner/cluster.rs
@@ -1,0 +1,247 @@
+// Copyright Materialize, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+use std::collections::BTreeSet;
+
+use mz_catalog::memory::objects::ClusterVariantManaged;
+use mz_controller::clusters::ReplicaLogging;
+use mz_controller_types::DEFAULT_REPLICA_LOGGING_INTERVAL;
+use mz_ore::instrument;
+use mz_sql::catalog::ObjectType;
+use mz_sql::plan::{AlterClusterPlan, AlterOptionParameter};
+use mz_sql::{plan, session::metadata::SessionMetadata};
+use tracing::Span;
+
+use crate::catalog::ReplicaCreateDropReason;
+use crate::{
+    coord::{AlterCluster, ClusterStage, Coordinator, Message, PlanValidity, StageResult, Staged},
+    session::Session,
+    AdapterError, ExecuteContext, ExecuteResponse,
+};
+
+use super::return_if_err;
+
+impl Staged for ClusterStage {
+    fn validity(&mut self) -> &mut PlanValidity {
+        match self {
+            Self::Alter(stage) => &mut stage.validity,
+        }
+    }
+
+    async fn stage(
+        self,
+        coord: &mut crate::coord::Coordinator,
+        ctx: &mut crate::ExecuteContext,
+    ) -> Result<crate::coord::StageResult<Box<Self>>, crate::AdapterError> {
+        match self {
+            Self::Alter(stage) => {
+                coord
+                    .sequence_alter_cluster(ctx.session(), stage.plan.clone())
+                    .await
+            }
+        }
+    }
+
+    fn message(self, ctx: crate::ExecuteContext, span: tracing::Span) -> crate::coord::Message {
+        Message::ClusterStageReady {
+            ctx,
+            span,
+            stage: self,
+        }
+    }
+
+    fn cancel_enabled(&self) -> bool {
+        // Cluster create and alter are not yet cancelable
+        false
+    }
+}
+
+impl Coordinator {
+    #[instrument]
+    pub(crate) async fn sequence_alter_cluster_staged(
+        &mut self,
+        ctx: ExecuteContext,
+        plan: plan::AlterClusterPlan,
+    ) {
+        let stage = return_if_err!(self.alter_cluster_validate(ctx.session(), plan).await, ctx);
+        self.sequence_staged(ctx, Span::current(), stage).await;
+    }
+
+    #[instrument]
+    async fn alter_cluster_validate(
+        &mut self,
+        session: &Session,
+        plan: plan::AlterClusterPlan,
+    ) -> Result<ClusterStage, AdapterError> {
+        let validity = PlanValidity {
+            transient_revision: self.catalog().transient_revision(),
+            dependency_ids: BTreeSet::new(),
+            cluster_id: Some(plan.id.clone()),
+            replica_id: None,
+            role_metadata: session.role_metadata().clone(),
+        };
+        Ok(ClusterStage::Alter(AlterCluster { validity, plan }))
+    }
+
+    pub(super) async fn sequence_alter_cluster(
+        &mut self,
+        session: &Session,
+        AlterClusterPlan {
+            id: cluster_id,
+            name: _,
+            options,
+        }: AlterClusterPlan,
+    ) -> Result<StageResult<Box<ClusterStage>>, AdapterError> {
+        use mz_catalog::memory::objects::ClusterVariant::*;
+
+        let config = self.catalog.get_cluster(cluster_id).config.clone();
+        let mut new_config = config.clone();
+
+        match (&new_config.variant, &options.managed) {
+            (Managed(_), AlterOptionParameter::Reset)
+            | (Managed(_), AlterOptionParameter::Unchanged)
+            | (Managed(_), AlterOptionParameter::Set(true)) => {}
+            (Managed(_), AlterOptionParameter::Set(false)) => new_config.variant = Unmanaged,
+            (Unmanaged, AlterOptionParameter::Unchanged)
+            | (Unmanaged, AlterOptionParameter::Set(false)) => {}
+            (Unmanaged, AlterOptionParameter::Reset)
+            | (Unmanaged, AlterOptionParameter::Set(true)) => {
+                // Generate a minimal correct configuration
+
+                // Size and disk adjusted later when sequencing the actual configuration change.
+                let size = "".to_string();
+                let disk = false;
+                let logging = ReplicaLogging {
+                    log_logging: false,
+                    interval: Some(DEFAULT_REPLICA_LOGGING_INTERVAL),
+                };
+                new_config.variant = Managed(ClusterVariantManaged {
+                    size,
+                    availability_zones: Default::default(),
+                    logging,
+                    replication_factor: 1,
+                    disk,
+                    optimizer_feature_overrides: Default::default(),
+                    schedule: Default::default(),
+                });
+            }
+        }
+
+        match &mut new_config.variant {
+            Managed(ClusterVariantManaged {
+                size,
+                availability_zones,
+                logging,
+                replication_factor,
+                disk,
+                optimizer_feature_overrides: _,
+                schedule,
+            }) => {
+                use AlterOptionParameter::*;
+                match &options.size {
+                    Set(s) => size.clone_from(s),
+                    Reset => coord_bail!("SIZE has no default value"),
+                    Unchanged => {}
+                }
+                match &options.disk {
+                    Set(d) => *disk = *d,
+                    Reset => *disk = self.catalog.system_config().disk_cluster_replicas_default(),
+                    Unchanged => {}
+                }
+                match &options.availability_zones {
+                    Set(az) => availability_zones.clone_from(az),
+                    Reset => *availability_zones = Default::default(),
+                    Unchanged => {}
+                }
+                match &options.introspection_debugging {
+                    Set(id) => logging.log_logging = *id,
+                    Reset => logging.log_logging = false,
+                    Unchanged => {}
+                }
+                match &options.introspection_interval {
+                    Set(ii) => logging.interval = ii.0,
+                    Reset => logging.interval = Some(DEFAULT_REPLICA_LOGGING_INTERVAL),
+                    Unchanged => {}
+                }
+                match &options.replication_factor {
+                    Set(rf) => *replication_factor = *rf,
+                    Reset => *replication_factor = 1,
+                    Unchanged => {}
+                }
+                match &options.schedule {
+                    Set(new_schedule) => {
+                        *schedule = new_schedule.clone();
+                    }
+                    Reset => *schedule = Default::default(),
+                    Unchanged => {}
+                }
+                if !matches!(options.replicas, Unchanged) {
+                    coord_bail!("Cannot change REPLICAS of managed clusters");
+                }
+            }
+            Unmanaged => {
+                use AlterOptionParameter::*;
+                if !matches!(options.size, Unchanged) {
+                    coord_bail!("Cannot change SIZE of unmanaged clusters");
+                }
+                if !matches!(options.availability_zones, Unchanged) {
+                    coord_bail!("Cannot change AVAILABILITY ZONES of unmanaged clusters");
+                }
+                if !matches!(options.introspection_debugging, Unchanged) {
+                    coord_bail!("Cannot change INTROSPECTION DEGUBBING of unmanaged clusters");
+                }
+                if !matches!(options.introspection_interval, Unchanged) {
+                    coord_bail!("Cannot change INTROSPECTION INTERVAL of unmanaged clusters");
+                }
+                if !matches!(options.replication_factor, Unchanged) {
+                    coord_bail!("Cannot change REPLICATION FACTOR of unmanaged clusters");
+                }
+            }
+        }
+
+        if new_config == config {
+            return Ok(StageResult::Response(ExecuteResponse::AlteredObject(
+                ObjectType::Cluster,
+            )));
+        }
+
+        match (&config.variant, new_config.variant) {
+            (Managed(config), Managed(new_config)) => {
+                self.sequence_alter_cluster_managed_to_managed(
+                    Some(session),
+                    cluster_id,
+                    config,
+                    new_config,
+                    ReplicaCreateDropReason::Manual,
+                )
+                .await?;
+            }
+            (Unmanaged, Managed(new_config)) => {
+                self.sequence_alter_cluster_unmanaged_to_managed(
+                    session, cluster_id, new_config, options,
+                )
+                .await?;
+            }
+            (Managed(_), Unmanaged) => {
+                self.sequence_alter_cluster_managed_to_unmanaged(session, cluster_id)
+                    .await?;
+            }
+            (Unmanaged, Unmanaged) => {
+                self.sequence_alter_cluster_unmanaged_to_unmanaged(
+                    session,
+                    cluster_id,
+                    options.replicas,
+                )?;
+            }
+        }
+        Ok(StageResult::Response(ExecuteResponse::AlteredObject(
+            ObjectType::Cluster,
+        )))
+    }
+}
